refactored to wrap historydetails page in Suspense for vercel
